services/requirements-dev.txt
-r requirements.txt
ruff>=0.7
pytest>=8
pytest-asyncio>=0.23
pytest-aiohttp>=1.0
